Removed last BillingHelper
