+++ src/AdminConsole/Services/ISharedBillingService.cs
@@ -1,3 +1,4 @@
+using Passwordless.AdminConsole.Models;
 using Stripe;
 
 namespace Passwordless.AdminConsole.Services;
@@ -9,6 +10,8 @@ public interface ISharedBillingService
     /// </summary>
     /// <returns></returns>
     Task UpdateUsageAsync();
+
+    Task<IReadOnlyCollection<PaymentMethodModel>> GetPaymentMethods(string? organizationBillingCustomerId);
 
     /// <summary>
     /// When a customer subscribes for the first time.
@@ -61,4 +64,8 @@ public interface ISharedBillingService
         string planName,
         string successUrl,
         string cancelUrl);
+
+    Task<string?> GetRedirectToUpgradeOrganization(string selectedPlan);
+    Task<string?> ChangePlanAsync(string app, string selectedPlan);
+    Task<(string subscriptionItemId, string priceId)> CreateSubscriptionItem(Organization org, string planSKU);
 }
